More STFReader changes.
Unified UNITS parsing adding the ability to limit the units according to context.

UNITS.None works faster than UNITS.Any, but is less tolerant, so as a first step, this is the most tolerant code, so we can find bugs.  Then we can progressively speed up the code by changing to only accept the appropriate units in their context.

git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@421 a9f3cfff-6c41-7446-b0ca-052bd761e6eb