Initial part of threading checks for OR:  - ThreadNameAttribute can be set on any method and will set the thread on which to assume it runs. Only one thread may be specified for any given method.  - CallOnThreadAttribute can be set on any class, constructor or method to say it is allowed to be called on a given thread. May be specified multiple times. Setting on a class defines allowed threads for all methods without their own attributes; method attributes completely replace class attributes. Overriding a method will inherit the base method's threading attributes and trying to redefine them is an error.
The FxCop rules I've written (not public just yet) check that all callers of methods with CallOnThread set are allowed by the threading attributes; i.e. callee's allowed threads must be the same or a superset of the sum of all calling threads. Methods without CallOnThread set are not checked in any way.

git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@482 a9f3cfff-6c41-7446-b0ca-052bd761e6eb