Added new ORTSPantographs block
For now, you can only use it to define a delay before the power is available on the pantograph (because there is no way to determine this delay with the pantograph animations).

Example : Pantograph 1 needs 2.25 secs to raise (and lower), pantograph 2 needs 3 secs.
ORTSPantographs(
    Pantograph(
        Delay( 2.25s )
    )
    Pantograph(
        Delay( 3s )
    )
)

Adding a third or a fourth pantograph will be useless for the moment because the rest of the simulator doesn't handle this case.

git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@2346 a9f3cfff-6c41-7446-b0ca-052bd761e6eb