http://www.elvastower.com/forums/index.php?/topic/29802-rolling-stock/page__st__30#entry238623
The minimum bogie count that should be officially recognized and was not supported is 3 since I did test a locomotive with 3 bogies and there was a comment from a post that mentioned MSTS does support 3 bogies.  I then added one more to make it 4 since doing so does not require major code change.

The default axles tested for and added is currently 3 per bogie, but the last test will add any additional axles not specifically tested for.

git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@4159 a9f3cfff-6c41-7446-b0ca-052bd761e6eb