Enhancements for OpenRailway Map:
- added web browser tab Open Rails icon (favicon)
- changed locomotive icon, now three icons: steam, electric, diesel
- added locomotive direction arrow
- added three OpenRailway layers to choose from: maximum speed, guage and electrification. Choice stored in registry and used the next opening of the map.
- added three layers with Open Rails data: track, named and rest. Choice also stored.

Don't know if these layers are usefull, anyway they can be switched on and off again.

Names of the layers are not translateble, is that a problem? Would be hard to implement.

For some routes the track layer takes too much performance from the web browser because of tracks size (for instance the MECoast line). But it gives you an idea of the layout of the track in the real world.

On the Ruebelandbahn for the "RLB Schneepflug Fahrbar" the locomotive arrow on the map is in the wrong direction (180 degrees). I can't find out why. Not seen this for any other locomotive.

(cherry picked from commit 5d8ea22f3abd2f0edf70cf86a66b3f7aa3f1d8ae)